add new document constructors replacing Head and Attach
