+++ apps/dart_quotes_server/lib/web/pages/quote_page.dart
@@ -21,7 +21,7 @@ class QuotePage extends AsyncStatelessComponent {
       return;
     }
 
-    yield Head(
+    yield Document.head(
       title: quote.quote,
       meta: {"description": '"${quote.quote}" - ${quote.author}'},
       children: [
