+++ packages/jaspr/lib/server.dart
@@ -2,11 +2,11 @@ library server;
 
 export 'package:shelf/shelf.dart' show Handler, Request, Response;
 
-export 'jaspr.dart' hide runApp;
+export 'jaspr.dart' hide runApp, Document;
+export 'src/components/document/document_server.dart';
 export 'src/server/adapters/element_boundary_adapter.dart';
 export 'src/server/adapters/head_scope_adapter.dart';
 export 'src/server/components/async.dart';
-export 'src/server/document.dart';
 export 'src/server/markup_render_object.dart';
 export 'src/server/run_app.dart';
 export 'src/server/server_app.dart';
