+++ packages/jaspr/lib/src/components/document.dart
@@ -1,0 +1,1 @@
+export 'document/document_client.dart' if (dart.library.io) 'document/document_server.dart' show Document;
