+++ packages/jaspr/test/browser/head/head_app.dart
@@ -3,7 +3,7 @@ import 'package:jaspr/jaspr.dart';
 class App extends StatelessComponent {
   @override
   Iterable<Component> build(BuildContext context) sync* {
-    yield Head(title: 'a', meta: {'test': 'b', 'c': 'd'});
+    yield Document.head(title: 'a', meta: {'test': 'b', 'c': 'd'});
     yield Page();
   }
 }
@@ -22,10 +22,10 @@ class _PageState extends State<Page> {
   Iterable<Component> build(BuildContext context) sync* {
     yield div([
       if (!pressed) ...[
-        Head(title: 'b', meta: {'c': 'e'}),
-        Head(title: 'c'),
+        Document.head(title: 'b', meta: {'c': 'e'}),
+        Document.head(title: 'c'),
       ] else ...[
-        Head(title: 'd'),
+        Document.head(title: 'd'),
       ]
     ]);
     yield button(onClick: () {
