+++ packages/jaspr/test/browser/head/head_browser_test.dart
@@ -2,7 +2,7 @@
 
 import 'dart:html';
 
-import 'package:jaspr/src/components/head/head_client.dart';
+import 'package:jaspr/src/components/document/document_client.dart';
 import 'package:jaspr_test/browser_test.dart';
 
 import 'head_app.dart';
@@ -12,7 +12,7 @@ void main() {
     testBrowser('should serve component', (tester) async {
       await tester.pumpComponent(App());
 
-      var nodes = HeadAdapter.instance.liveHeadNodes.toList();
+      var nodes = AttachAdapter.instanceFor(AttachTarget.head).liveNodes.toList();
 
       expect(nodes, hasLength(3));
       expect((nodes[0] as Element).outerHtml, equals('<title>c</title>'));
@@ -21,7 +21,7 @@ void main() {
 
       await tester.click(find.tag('button'));
 
-      nodes = HeadAdapter.instance.liveHeadNodes.toList();
+      nodes = AttachAdapter.instanceFor(AttachTarget.head).liveNodes.toList();
 
       expect((nodes[0] as Element).outerHtml, equals('<title>d</title>'));
       expect((nodes[1] as Element).outerHtml, equals('<meta name="test" content="b">'));
