+++ packages/jaspr/test/server/head/head_app.dart
@@ -3,10 +3,10 @@ import 'package:jaspr/jaspr.dart';
 class App extends StatelessComponent {
   @override
   Iterable<Component> build(BuildContext context) sync* {
-    yield Head(title: 'a', meta: {'test': 'b', 'c': 'd'});
+    yield Document.head(title: 'a', meta: {'test': 'b', 'c': 'd'});
     yield div([
-      Head(title: 'b', meta: {'c': 'e'}),
-      Head(title: 'c'),
+      Document.head(title: 'b', meta: {'c': 'e'}),
+      Document.head(title: 'c'),
     ]);
   }
 }
